Small bug fix: LiftSparkMaxPort type changed from double to int

+++ src/main/java/frc/robot/Constants.java
@@ -44,7 +44,7 @@ public final class Constants {
     public static final double kDInternalElevator = 0.001;
 
 
-    public static final double kLiftSparkMaxPort = 16;
+    public static final int    kLiftSparkMaxPort = 16;
     public static final double kMaxLiftSpeed = 0.2;
     public static final double kMinLiftSpeed = 0.1;
     public static final double kLiftConversionFactor = 0.0;
